Create RailsCasts with basic functionality
At this point RailsCasts has the ability to take a url and return an
array of Episodes.  To complete this functionality an Episodes and a
NullNokogiri class were added.

Episodes is a class to handle the data structure for an individual
episode.

NullNokogiri uses the Null Object Pattern to handle cases were Nokogiri
returns no results.

Specs are provided for each of the methods in RailsCasts.  In this
commit all methods are private, so the specs utilize "send" to send
messages to the private methods.  This should be changed so that the
testing using the publicly available attr_readers.