Update NullWebsite to display cli message
Prior to this commit, NullWebsite only handled cases where a website
couldn't be opened.  With this commit, NullWebsite now handles cases
were a website can't be opened or a blank filter is provided.

To make this change a bug was fixed in RailsCasts#initialize where the
call to Website.get_list had a hard coded argument of '.episode' for the
filter parameter.  This was changed to passing it the filter argument
from initialize.

Fixing that bug required adding NullWebsite#collect.  #collect returns
self (an instance of NullWebsite) so that additional error handling can
be controlled at other points.

The final fix was the addition of NullWebsite#each_with_object.  This
method handles the part of the cli that displays a list of episodes.
The method calls CLI.display with the message 'No Episodes Found'.