Fix CLI interface looping bug
The CLI interface had a looping issue where the user needed to type back
or exit several times for the command to work.  This code fixes that
problem by having a loop for both the list view and the detail view.

This commit also adds the open functionality that will open an episode
in the user's web browser.