Refactored cli checks for episode number
This commit finishes refactoring the cli's checks for a valid episode
number selection.  The tests were moved from CLI and the
CLI.valid_episode_number to RetroCasts.interger? and
RailsCasts#episode?.

This interface seems cleaner, since a site should be able to figure out
if it has an episode.

Also added RailsCasts#episode to get an episode based on a 1 index.