Adds search and page to RailsCasts#new
This commit modifies RailsCasts#new so that RailsCasts can be
instantiated with optional search and page parameters.  Additionally, a
private #url_builder method was created to handle building urls need for
the RailsCasts website.

This interface allows a new instance of RailsCasts to be created
whenever the url needs to be changed, i.e. to search or find another
page.