Add site caching
Prior to this commit, each website page load was completed on demand.
For instance, starting the program, going to the next page, and then
going back a page required 3 requests to railcasts.com.  This was
somewhat slow and inefficient.

Since, RailsCasts creates news instances each time new episodes are
requested, a cache could not be stored in that class.  Because, the
Website class already acts as a wrapper for website requests, it seems
like the best place to also handle caching webpages for future use.

To accomplish this, Website was changed from a module to a class and
than changed to a class with instance methods instead of only class
methods.  An initial take on caching tried to use on class methods with
a class variable for caching, but this caused strange issues in the test
suite.

RailsCasts was modified to take an instance of Website on instantiation
with a default, new instance of Website.  This way in the RailsCasts
methods that create new instances of RailsCasts, the existing instance
of website could be injected.