Refactor RailsCasts state changes
State changes in RailsCasts happen with three methods: #get_search,
 #next_page, and #prev_page.  When these methods are called @url and
@episodes need to change.  Prior to the commit, these state changes were
completed by returning a new instance of RailsCasts.  One reason for
this decision was to utilize the logic already contained in #initialize.

With this commit, state changes are handled by #update.  Update just
contains the relevant code to get a new page from Website and have the
new episodes parsed.  Instead of calling #initialize, the state change
methods now simply change instance variables as needed and then run
update to complete the refresh.

This change is more consistent with object oriented conventions and it
also simplifies the code in RetroCasts.start.